dai: support multiple configurations per dai
The dai configuration is currently not being kept by the
dai layer. When received from the IPC it is instead directly
passed to the underlying driver. If the Linux driver sends
multiple configurations for a dai, only the last one would
be kept.

This patch adds the necessary changes to keep multiple configs
as part of the dai.

Signed-off-by: Dragos Tarcatu <dragos_tarcatu@mentor.com>